Bug 1316083 - Refactor Localization into DOMLocalization. r=gandalf
- The old LocalizationObserver class is no more.  Instead, there are two
subclasses of Localization: DOMLocalization and DocumentLocalization.

- DOMLocalization is a subclass of Localization and implements methods for
  translating DOM trees and elements, like translateFragment and
  translateElement.  Localization itself is environment-agnostic.

- DocumentLocalization is a subclass of DOMLocalization.  In HTML/XUL
  documents, an instance of DocumentLocalization will be available as
  document.l10n.  It implements the observer interface and supports delegating
  to sub-localizations. The sub-localizations are stored in the
  document.l10n.delegates map.

- I considered factoring the observer interface into its own class or a mixin,
  but eventually decided to leave it as part of DocumentLocalization to keep
  things simple.

- We used to have two methods related to observing roots: observeRoot and
  disconnectRoot.  They would serve two purposes:

    1) observe/unobserve the root via the MutationObserver instance, and
    2) add/remove the root to/from the list of known roots used by top-down
       methods like translateRoots.

  In this PR, I separated these two purposes into two pairs of methods:

    - observeRoot and unobserveRoot are available only on DocumentLocalization
      (because it implements the observer interface) and they only handle 1)
      from above; they aren't part of the public API,

    - connectRoot and disconnectRoot are available on every DOMLocalization
      (because now every DOMLocalization including DocumentLocalization has its
      own set of roots) and it handles 2) from above; it will also use
      observeRoot/unobserveRoot if they're available;  thus, connectRoot and
      disconnectRoot can be considered part of the public API.
