+++ examples/web.html
@@ -19,6 +19,6 @@
     <p data-l10n-id="unreadEmails" data-l10n-args='{"emailCount": 5}'></p>
     <input type="text" data-l10n-id="input" />
     <div data-l10n-id="checkbox"><input type="checkbox"></div>
-    <h2 data-l10n-id="extra-foo" data-l10n-bundle="extra"></h2>
+    <h2 data-l10n-id="extra-foo" data-l10n-with="extra"></h2>
   </body>
 </html>
